refactor(SwiftInterface): give the pack-generic type the evolution builder's primary name
The user's round-two decision meant the pack on the TYPE, not a pack
initializer on a non-generic class. Two probe facts shape the split:
packs in a type's generic parameter list require the Swift 5.9 runtime
(@available macOS 14+, compiler-enforced) while packs in function
position need no gate, and same-element requirements
(repeat each MachO == M) are not yet supported — so a runtime-count
[MachO] initializer cannot live on the pack type, and a pack's
compile-time arity cannot serve the CLI's / a host's runtime-N axis at
all.

Hence two public types: AnySwiftEvolutionInterfaceBuilder, the
type-erased runtime-N workhorse (the former class renamed following the
AnyView/AnySequence erasure convention; both its initializers available
everywhere; the CLI now uses it), and
SwiftEvolutionInterfaceBuilder<each MachO: FieldLayoutRenderable>, the
availability-gated pack-generic facade for compile-time-fixed axes —
constructs-then-erases, behavior byte-identical by construction, pinned
by packGenericFacadeMatchesTheErasedBuilder.
