+++ AGENTS.md
@@ -139,7 +139,7 @@ The interface generation is split into layered peer modules over a shared `Swift
 **SwiftInterface** - Thin orchestrator tying indexing + printing into a full interface dump
 - `SwiftInterfaceBuilder` - Main builder, call `prepare()` then `printRoot()`
 - `.swiftinterface` file types (`SwiftInterfaceFile`, `SwiftInterfaceParser`, …)
-- `SwiftEvolutionInterfaceBuilder` - N ≥ 2 versions rendered as ONE **union interface with lifecycle annotations** (evolution proposal draft-swift-evolution-interface-builder; the N-way analogue of the `SwiftDiffableInterfaceBuilder`+`Renderer` pair). Division of labor is the load-bearing rule: annotation facts come SOLELY from `ABIEvolutionBuilder` over the versions' frozen snapshots (`EvolutionAnnotationIndex` joins on the same `ABIKey`/`MemberRecord` constructions `ABIDiffer` freezes — a lookup miss IS the "present throughout, never changed" verdict and renders bare), while text renders from the live models: each declaration from the LAST version that has it, so a modified member shows one line (newest generation) with the old shape in the phrase (`modified in 26.0: old → new`; an identical arrow collapses to the bare phrase). Union order = newest version's order as spine, absent declarations appended in their last-carrying version's order (`matchAcrossVersions`, the N-way `matchByKey`). Format layer (`EvolutionMarking`): trailing `// [●●○] removed in 26.0` comments (bitmap + phrases, legend header mapping positions to labels), per-block column alignment capped at 72 (overflow → own line one level deeper), annotation attached to a unit's LAST line, reporter-mirrored warnings tail. The public class is deliberately NON-generic — each version erases at construction (`EvolutionVersionRendering`/`EvolutionVersionUnit`), so the everywhere-available homogeneous `[MachO]` init and the parameter-pack heterogeneous init (`@available(macOS 14…)` — packs need the Swift 5.9 runtime) produce the same type. All inputs must be binaries (snapshots carry no renderable interface); protocols' stripped `pwtslot:` records are not rendered (no declaration — same as `diff --interface`). CLI: `swift-section evolution --interface` (mutually exclusive with `--json`/`--summary-only`; colorizes by event kind at the CLI layer). Structured stream for hosts: `@_spi(Support) annotatedBlocks()` → `[[EvolutionLine]]`
+- `SwiftEvolutionInterfaceBuilder` - N ≥ 2 versions rendered as ONE **union interface with lifecycle annotations** (evolution proposal draft-swift-evolution-interface-builder; the N-way analogue of the `SwiftDiffableInterfaceBuilder`+`Renderer` pair). Division of labor is the load-bearing rule: annotation facts come SOLELY from `ABIEvolutionBuilder` over the versions' frozen snapshots (`EvolutionAnnotationIndex` joins on the same `ABIKey`/`MemberRecord` constructions `ABIDiffer` freezes — a lookup miss IS the "present throughout, never changed" verdict and renders bare), while text renders from the live models: each declaration from the LAST version that has it, so a modified member shows one line (newest generation) with the old shape in the phrase (`modified in 26.0: old → new`; an identical arrow collapses to the bare phrase). Union order = newest version's order as spine, absent declarations appended in their last-carrying version's order (`matchAcrossVersions`, the N-way `matchByKey`). Format layer (`EvolutionMarking`): trailing `// [●●○] removed in 26.0` comments (bitmap + phrases, legend header mapping positions to labels), per-block column alignment capped at 72 (overflow → own line one level deeper), annotation attached to a unit's LAST line, reporter-mirrored warnings tail. The public surface is TWO types: `AnySwiftEvolutionInterfaceBuilder`, the type-ERASED runtime-N workhorse (each version erases at construction via `EvolutionVersionRendering`/`EvolutionVersionUnit`; homogeneous `[MachO]` init + heterogeneous pack init, both available everywhere — packs in *function* position need no availability gate), and `SwiftEvolutionInterfaceBuilder<each MachO: FieldLayoutRenderable>`, the pack-generic façade for compile-time-fixed axes (`@available(macOS 14…)` — packs in a TYPE's generic parameter list DO need the Swift 5.9 runtime; constructs-then-erases, behavior byte-identical by construction, pinned by `packGenericFacadeMatchesTheErasedBuilder`). A pack's arity is compile-time, so runtime-N (the CLI, RuntimeViewer's user-picked versions) always goes through the erased type; same-element requirements (`repeat each MachO == M`) are not yet supported by the toolchain, so the array init cannot live on the pack type. All inputs must be binaries (snapshots carry no renderable interface); protocols' stripped `pwtslot:` records are not rendered (no declaration — same as `diff --interface`). CLI: `swift-section evolution --interface` (mutually exclusive with `--json`/`--summary-only`; colorizes by event kind at the CLI layer). Structured stream for hosts: `@_spi(Support) annotatedBlocks()` → `[[EvolutionLine]]`
 - `SwiftInterfaceBuilderOpaqueTypeProvider` - Expands `some` return types from the opaque type descriptor's generic requirements. Primary-associated-type arguments are **attributed per protocol** (evolution proposal 0006), never distributed blindly across the composition: a same-type constraint attaches to the protocol its **anchor** names (the associated type's declaring protocol, kept by the demangler in `dependentAssociatedTypeRef` — identity comparison only, so it works offline via bind symbols), to a protocol whose refine closure contains the anchor, or — for compiler-collapsed equivalence classes like `TestCollection<[A]>` — by name fallback (only when no anchor matched, the protocol itself declares that associated-type name, the candidate is unique, and the anchor lies **outside** the composition; a collapsed pin and a never-pinned same-named member are byte-identical in the descriptor, so an in-composition anchor forbids the fallback). Protocol facts (refine signature + associated-type names) resolve through `ProtocolFactsResolver`: a reachable descriptor (same image on both readers; **any** image in-process — `MachOImage` dereferences the requirement's indirect pointer cross-image) merged with `BuiltinStandardLibraryProtocolFacts` (the only source of primary names/order — SE-0346 leaves no runtime trace — and the offline fallback for bind-only externals). Unknown facts attach nothing (a missed parameter beats a fabricated one), so `MachOFile` and `MachOImage` output depth may differ — accepted, in-process is strictly additive. Composition order is the descriptor's canonical order; source order is not recoverable. See [Documentations/Internal/OpaquePrimaryAssociatedTypeAttribution.md](Documentations/Internal/OpaquePrimaryAssociatedTypeAttribution.md) (implementation note) and [Documentations/Internal/OpaqueReturnTypeResolution.md](Documentations/Internal/OpaqueReturnTypeResolution.md) (domain deep-dive: descriptor encoding, anchor/collapse mechanics, byte-level debugging).
 
 Printing and indexing are peers — neither depends on the other.
