+++ Sources/SwiftInterface/SwiftEvolutionInterfaceBuilder.swift
@@ -1,164 +1,75 @@
 import Foundation
 import SwiftDeclaration
-@_spi(Support) import SwiftIndexing
+import SwiftIndexing
 import SwiftDeclarationRendering
 import SwiftDiffing
 import MachOSwiftSection
 import Semantic
-import SwiftStdlibToolbox
 
-/// Renders one module's ABI across N ≥ 2 ordered binary versions as a single
-/// **union interface with lifecycle annotations** — the N-way, human-readable
-/// counterpart of `swift-section evolution`'s lineage report, and the
-/// evolution analogue of the `SwiftDiffableInterfaceBuilder` +
-/// `SwiftDiffableInterfaceRenderer` pair.
+/// The pack-generic evolution interface builder: the version axis' reader
+/// types live in the type's own signature —
+/// `SwiftEvolutionInterfaceBuilder<MachOFile, MachOFile, MachOFile>`, or a
+/// heterogeneous mix like `<MachOFile, MachOImage>` — for callers whose
+/// version list is fixed at compile time.
 ///
-/// Every declaration that ever existed on the axis renders exactly once, from
-/// the last version that has it; declarations that changed carry a trailing
-/// `// [●●○] removed in 26.0`-style annotation (presence bitmap + event
-/// phrases, legend in the file header), and declarations present throughout
-/// with no changes carry none. Annotation facts come from `ABIEvolutionBuilder`
-/// over the same frozen snapshots the lineage report uses, so the two views
-/// cannot disagree.
+/// A thin façade over ``AnySwiftEvolutionInterfaceBuilder``: construction
+/// erases the pack immediately, so behavior is identical by construction —
+/// there is exactly one rendering pipeline. Two consequences of the pack
+/// living in *type* position:
 ///
-/// Usage: construct with N binaries (oldest first) and their axis labels,
-/// `try await prepare()`, then `printAnnotatedInterface()`. All inputs must be
-/// live binaries — a persisted snapshot carries no renderable interface.
+/// - **Availability**: parameter packs in generic types need the Swift 5.9
+///   runtime (`parameter packs in generic types are only available in
+///   macOS 14.0.0 or newer`), hence the gate — packs in *function* position
+///   (the erased builder's heterogeneous initializer) need none.
+/// - **Arity is compile-time**: a pack cannot express "N versions decided at
+///   runtime". The CLI and any host letting the user pick an arbitrary set of
+///   versions use ``AnySwiftEvolutionInterfaceBuilder`` instead.
 ///
-/// The class is deliberately **non-generic**: each version is erased at
-/// construction, so the homogeneous-array initializer (all versions the same
-/// reader type — the CLI's N `MachOFile`s) and the parameter-pack initializer
-/// (heterogeneous reader types, gated on the Swift 5.9 runtime that parameter
-/// packs require) both produce the same type.
-public final class SwiftEvolutionInterfaceBuilder: Sendable {
-    /// Oldest → newest, index-aligned with `labels`.
-    private let versionUnits: [any EvolutionVersionRendering]
+/// See ``AnySwiftEvolutionInterfaceBuilder`` for the pipeline's semantics
+/// (union interface, lifecycle annotations, annotation-fact sourcing).
+@available(macOS 14.0, iOS 17.0, tvOS 17.0, watchOS 10.0, *)
+public final class SwiftEvolutionInterfaceBuilder<each MachO: FieldLayoutRenderable>: Sendable {
+    /// The type-erased core every entry point delegates to. Exposed so a
+    /// pack-typed builder can be handed to API that takes the erased form.
+    public let erased: AnySwiftEvolutionInterfaceBuilder
 
-    /// The version-axis labels, oldest first (e.g. `["17.0", "18.0", "26.0"]`).
-    public let labels: [String]
-
-    @Mutex
-    private var preparedEvolution: ABIEvolution? = nil
-
-    /// Homogeneous construction: N versions of the same reader type.
-    /// Available on every platform this package supports.
+    /// One version per pack element, oldest first, labels index-aligned.
     ///
     /// - Throws: `ABIEvolutionError.fewerThanTwoVersions` /
     ///   `.labelCountMismatch` on invalid input shapes.
-    public init<MachO: FieldLayoutRenderable>(
+    public init(
         configuration: SwiftDeclarationIndexConfiguration = .init(),
         eventHandlers: [SwiftIndexEvents.Handler] = [],
-        versions: [MachO],
+        versions: repeat each MachO,
         labels: [String]
     ) throws {
-        try Self.validate(versionCount: versions.count, labelCount: labels.count)
-        self.versionUnits = versions.map {
-            EvolutionVersionUnit(configuration: configuration, eventHandlers: eventHandlers, machO: $0)
-        }
-        self.labels = labels
-    }
-
-    /// Heterogeneous construction: each version brings its own reader type
-    /// (e.g. a standalone `MachOFile` next to a dyld-cache image). Parameter
-    /// packs need the Swift 5.9 runtime (SE-0393), hence the availability
-    /// floor; on older systems use the homogeneous initializer.
-    @available(macOS 14.0, iOS 17.0, tvOS 17.0, watchOS 10.0, *)
-    public init<each Reader: FieldLayoutRenderable>(
-        configuration: SwiftDeclarationIndexConfiguration = .init(),
-        eventHandlers: [SwiftIndexEvents.Handler] = [],
-        versions: repeat each Reader,
-        labels: [String]
-    ) throws {
-        var units: [any EvolutionVersionRendering] = []
-        for machO in repeat each versions {
-            units.append(EvolutionVersionUnit(configuration: configuration, eventHandlers: eventHandlers, machO: machO))
-        }
-        try Self.validate(versionCount: units.count, labelCount: labels.count)
-        self.versionUnits = units
-        self.labels = labels
-    }
-
-    private static func validate(versionCount: Int, labelCount: Int) throws {
-        guard versionCount >= 2 else {
-            throw ABIEvolutionError.fewerThanTwoVersions(versionCount: versionCount)
-        }
-        guard labelCount == versionCount else {
-            throw ABIEvolutionError.labelCountMismatch(labelCount: labelCount, versionCount: versionCount)
-        }
-    }
-
-    // MARK: - Preparation
-
-    /// Indexes every version (full member indexing included), freezes each
-    /// into a snapshot, and builds the `ABIEvolution` annotation matrix. Must
-    /// complete before any rendering entry point.
-    public func prepare() async throws {
-        for versionUnit in versionUnits {
-            try await versionUnit.prepare()
-        }
-        let snapshots = versionUnits.map { $0.snapshot() }
-        let versionDescriptors = labels.map { ABIVersionDescriptor(label: $0) }
-        preparedEvolution = try ABIEvolutionBuilder().evolution(of: snapshots, versions: versionDescriptors)
-    }
-
-    /// The evolution built by `prepare()` — the same value the lineage report
-    /// and JSON paths consume, exposed so a host can reuse it directly (CI
-    /// verdicts via `hasBreakingChange`, warnings display). `nil` until
-    /// `prepare()` completes.
-    public var evolution: ABIEvolution? { preparedEvolution }
-
-    // MARK: - Rendering
-
-    /// The full annotated union interface: legend header, every declaration
-    /// block, and a warnings tail when the evolution carries diagnostics.
-    ///
-    /// - Throws: ``SwiftEvolutionInterfaceBuilderError/notPrepared`` when
-    ///   called before `prepare()`.
-    public func printAnnotatedInterface() async throws -> SemanticString {
-        let evolution = try requirePrepared()
-        let renderer = SwiftEvolutionInterfaceRenderer(
-            versions: versionUnits,
-            annotations: EvolutionAnnotationIndex(evolution: evolution)
+        self.erased = try AnySwiftEvolutionInterfaceBuilder(
+            configuration: configuration,
+            eventHandlers: eventHandlers,
+            versions: repeat each versions,
+            labels: labels
         )
-        let blocks = await renderer.annotatedBlocks()
-        return EvolutionMarking.renderInterface(blocks: blocks, evolution: evolution)
     }
 
-    /// The structured line stream behind ``printAnnotatedInterface()``: the
-    /// outer array is the top-level declaration blocks in render order, each
-    /// inner array one block's lines with their lifecycle annotations still
-    /// data (not text). For hosts that render the evolution themselves
-    /// (custom coloring, folding) — the evolution analogue of the two-sided
-    /// renderer's `annotatedDiffBlocks()`.
+    /// The version-axis labels, oldest first.
+    public var labels: [String] { erased.labels }
+
+    /// See ``AnySwiftEvolutionInterfaceBuilder/evolution``.
+    public var evolution: ABIEvolution? { erased.evolution }
+
+    /// See ``AnySwiftEvolutionInterfaceBuilder/prepare()``.
+    public func prepare() async throws {
+        try await erased.prepare()
+    }
+
+    /// See ``AnySwiftEvolutionInterfaceBuilder/printAnnotatedInterface()``.
+    public func printAnnotatedInterface() async throws -> SemanticString {
+        try await erased.printAnnotatedInterface()
+    }
+
+    /// See ``AnySwiftEvolutionInterfaceBuilder/annotatedBlocks()``.
     @_spi(Support)
     public func annotatedBlocks() async throws -> [[EvolutionLine]] {
-        let evolution = try requirePrepared()
-        let renderer = SwiftEvolutionInterfaceRenderer(
-            versions: versionUnits,
-            annotations: EvolutionAnnotationIndex(evolution: evolution)
-        )
-        return await renderer.annotatedBlocks()
-    }
-
-    private func requirePrepared() throws -> ABIEvolution {
-        guard let evolution = preparedEvolution else {
-            throw SwiftEvolutionInterfaceBuilderError.notPrepared
-        }
-        return evolution
-    }
-}
-
-/// Usage-contract failures of `SwiftEvolutionInterfaceBuilder`. (Input-shape
-/// failures throw `ABIEvolutionError` from the initializers instead, shared
-/// with `ABIEvolutionBuilder`.)
-public enum SwiftEvolutionInterfaceBuilderError: Swift.Error, Equatable, CustomStringConvertible {
-    /// A rendering entry point ran before `prepare()` completed.
-    case notPrepared
-
-    public var description: String {
-        switch self {
-        case .notPrepared:
-            return "SwiftEvolutionInterfaceBuilder needs prepare() to complete before rendering."
-        }
+        try await erased.annotatedBlocks()
     }
 }
